Actualiza estilos y agrega elementos
