+++ index/comedor.html
@@ -1,0 +1,100 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>Comedor Universitario Neuquen</title>
+        <meta charset="utf-8">
+        <meta lang="es">
+        <link rel="stylesheet" href="index.css">
+        <!--Los siguientes links son de bootstrap-->
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
+        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.9.1/font/bootstrap-icons.css">
+    </head>
+    <body>
+        <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+            <a class="navbar-brand" id="title-Unco">COMEDOR UNIRVERSITARIO NEUQUEN</a>
+            <img class="logo-Unco" alt="Logo-UNCO" src = "https://web.curza.uncoma.edu.ar/images/logo_blanco.png">
+            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+              <span class="navbar-toggler-icon"></span>
+            </button>
+            <div class="collapse navbar-collapse" id="navbarSupportedContent">
+              <ul class="navbar-nav mr-auto">
+                <li class="nav-item active">
+                  <a class="nav-link" href="#">Ingresar</a>
+                </li>
+                <li class="nav-item">
+                  <a class="nav-link" href="#">Registrarse</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="#">Noticias</a>
+                </li>
+          </nav>
+        <div class="background-img-menu">
+            <div id="darken-background">
+                <h3 class="tip">¡Reserva tu menu para hoy y disfrutas de los mejores menus!</h3>
+                <div class="container-row">
+                    <img class="people-eat" alt="Siluetas Personas Comiendo" src="https://images.vexels.com/media/users/3/152385/isolated/preview/0e6b3970a47d518d95d940209c8bd694-pareja-sentada-en-la-mesa-de-comedor-silueta.png">;
+                    <img class="fork" alt="Siluetas Personas Comiendo" src="https://e7.pngegg.com/pngimages/879/351/png-clipart-cutlery-household-silver-fork-fork-plate-black-and-white.png">;
+                    <img class="fork" alt="Siluetas Personas Comiendo" src="https://img2.freepng.es/20180423/hcq/kisspng-chef-computer-icons-paella-cooking-male-chef-5ade1e1f7654c8.8565448715245061434847.jpg">;
+                </div>
+                <hr style="color: white;">
+                <h2 class="title-menu-today">Comida de Hoy</h2>
+                <div class="container-menus">
+                    <h3 id="time-foods">Desayuno - Horarios 9:30-11</h3>
+                    <div class="container-row">
+                        <button id="clickeable">
+                            <div class="little-box">
+                                <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
+                                <div class="little-box-title">
+                                    <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
+                                </div>
+                            </div>
+                        </button>
+                        <div class="little-box">
+                            <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
+                            <div id="clickeable" class="little-box-title">
+                                <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
+                            </div>
+                        </div>
+                    </div>
+                    <h3 id="time-foods">Almuerzo - Horarios 12-14</h3>
+                    <div class="container-row">
+                        <button id="clickeable">
+                            <div class="little-box">
+                                <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
+                                <div class="little-box-title">
+                                    <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
+                                </div>
+                            </div>
+                        </button>
+                        <div class="little-box">
+                            <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
+                            <div id="clickeable" class="little-box-title">
+                                <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
+                            </div>
+                        </div>
+                    </div>
+                    <h3 id="time-foods">Merienda - Horarios 17-19</h4>
+                        <div class="container-row">
+                            <button id="clickeable">
+                                <div class="little-box">
+                                    <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
+                                    <div class="little-box-title">
+                                        <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
+                                    </div>
+                                </div>
+                            </button>
+                            <div class="little-box">
+                                <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
+                                <div id="clickeable" class="little-box-title">
+                                    <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
+                                </div>
+                            </div>
+                        </div>
+                    <script type="module" src="index.js"></script>
+                </div>
+                <p class ="next-meals"><a href="index2.html">Proximas Comidas</a>
+            </div>
+        </div>
+          <footer class="contact">Datos de Contacto</div>
+    </body>
+</html>
