--- index/index.html
@@ -1,72 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Nombre de la pagina</title>
-        <meta charset="utf-8">
-        <meta lang="es">
-        <link rel="stylesheet" href="index.css">
-        <!--Los siguientes links son de bootstrap-->
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
-        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.9.1/font/bootstrap-icons.css">
-    </head>
-    <body>
-        <h1 class="conteiner">Comedor Universitario Neuquen</h1>
-        <h2 class="conteiner">Comida de Hoy</h2>
-        <div class="container-menus">
-            <h3 id="time-foods">Desayuno - Horarios 9:30-11</h3>
-            <div class="container-row">
-                <button id="clickeable">
-                    <div class="little-box">
-                        <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
-                        <div class="little-box-title">
-                            <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
-                        </div>
-                    </div>
-                </button>
-                <div class="little-box">
-                    <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
-                    <div id="clickeable" class="little-box-title">
-                        <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
-                    </div>
-                </div>
-            </div>
-            <h3 id="time-foods">Almuerzo - Horarios 12-14</h3>
-            <div class="container-row">
-                <button id="clickeable">
-                    <div class="little-box">
-                        <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
-                        <div class="little-box-title">
-                            <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
-                        </div>
-                    </div>
-                </button>
-                <div class="little-box">
-                    <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
-                    <div id="clickeable" class="little-box-title">
-                        <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
-                    </div>
-                </div>
-            </div>
-            <h3 id="time-foods">Merienda - Horarios 17-19</h4>
-                <div class="container-row">
-                    <button id="clickeable">
-                        <div class="little-box">
-                            <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
-                            <div class="little-box-title">
-                                <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
-                            </div>
-                        </div>
-                    </button>
-                    <div class="little-box">
-                        <img class="small-image" alt="Imagen del menu" src="https://assets.unileversolutions.com/recipes-v2/209413.jpg">
-                        <div id="clickeable" class="little-box-title">
-                            <h4 id="little-title-style">Et voluptate et nostrud eiusmod</h4>
-                        </div>
-                    </div>
-                </div>
-            <script type="module" src="index.js"></script>
-          </div>
-          <p class ="next-meals"><a href="index2.html">Proximas Comidas</a>
-          <footer class="contact">Datos de Contacto</div>
-    </body>
-</html>
